requirements_auto.txt
amqp==2.5.2
astroid==2.2.5
attrs==19.3.0
backcall==0.1.0
billiard==3.6.3.0
bleach==3.1.4
celery==4.4.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.2
django-extensions==2.2.9
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
idna==2.8
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.0
Jinja2==2.10.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
kombu==4.6.8
lazy-object-proxy==1.4.1
locustio==0.11.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msgpack==0.6.1
mysqlclient==1.4.2.post1
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
numpy==1.17.2
openpyxl==2.6.3
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
pycparser==2.19
Pygments==2.4.2
pylint==2.3.1
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2019.1
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.44.1
traitlets==4.3.2
transliterate==1.10.2
typed-ast==1.4.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
